% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcOLS.R
\name{olsFunc}
\alias{olsFunc}
\title{Function for OLS Estimators}
\usage{
olsFunc(formula, data = NULL, intercept = TRUE)
}
\arguments{
\item{formula}{An object of class \link[stats:formula]{stats::formula}}

\item{data}{an optional data frame, list or environment containing the variables in the model.}

\item{intercept}{logical expression whether an intercept should be included in the linear model.}
}
\value{
A numeric vector including the coefficient
}
\description{
Function for OLS Estimators
}
\examples{
n <- 100
X <- rnorm(n)
y <- 1.5*X+rnorm(n)
beta_OLS <- olsFunc(y~X)

n <- 100
X1 <- rnorm(n)
X2 <- rnorm(n)
y <- X1 + 1.5*X2+rnorm(n)
beta_OLS2 <- olsFunc(y~X1+X2)

df <- data.frame(y,X1, X2)
colnames(df) <- c("Y", "A", "B")
beta_OLS3 <- olsFunc(Y~A+B, data=df)
}
