% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcOLS.R
\name{really_old_olsFunc}
\alias{really_old_olsFunc}
\title{Deprecated function for OLS}
\usage{
really_old_olsFunc(y, X)
}
\arguments{
\item{y}{a vector of a dependent variable}

\item{X}{a matrix of numerical explanatory variables}
}
\value{
A vector of OLS estimates
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
